refactor: put pipeline operations in their own module
docs: added comments for functions
